move `TransientErrorHandler` under `can::` #verification #docs #sonar
