+++ test/unittest/transport/can/test_can_svc_tx_sessions.cpp
@@ -8,13 +8,14 @@
 #include "../../memory_resource_mock.hpp"
 #include "../../tracking_memory_resource.hpp"
 #include "../../virtual_time_scheduler.hpp"
+#include "can_gtest_helpers.hpp"
 #include "media_mock.hpp"
 
 #include <canard.h>
 #include <cetl/pf17/cetlpf.hpp>
+#include <libcyphal/transport/can/can_transport_impl.hpp>
 #include <libcyphal/transport/can/media.hpp>
 #include <libcyphal/transport/can/svc_tx_sessions.hpp>
-#include <libcyphal/transport/can/transport.hpp>
 #include <libcyphal/transport/errors.hpp>
 #include <libcyphal/transport/svc_sessions.hpp>
 #include <libcyphal/transport/types.hpp>
@@ -24,7 +25,6 @@
 #include <gtest/gtest.h>
 
 #include <array>
-#include <chrono>
 #include <utility>
 
 namespace
@@ -32,7 +32,8 @@ namespace
 
 using libcyphal::TimePoint;
 using libcyphal::UniquePtr;
-using namespace libcyphal::transport;  // NOLINT This our main concern here in the unit tests.
+using namespace libcyphal::transport;       // NOLINT This our main concern here in the unit tests.
+using namespace libcyphal::transport::can;  // NOLINT This our main concern here in the unit tests.
 
 using cetl::byte;
 
@@ -73,13 +74,13 @@ protected:
         return scheduler_.now();
     }
 
-    UniquePtr<can::ICanTransport> makeTransport(cetl::pmr::memory_resource& mr, const NodeId local_node_id)
+    UniquePtr<ICanTransport> makeTransport(cetl::pmr::memory_resource& mr, const NodeId local_node_id)
     {
-        std::array<can::IMedia*, 1> media_array{&media_mock_};
+        std::array<IMedia*, 1> media_array{&media_mock_};
 
         auto maybe_transport = can::makeTransport(mr, media_array, 16);
-        EXPECT_THAT(maybe_transport, VariantWith<UniquePtr<can::ICanTransport>>(NotNull()));
-        auto transport = cetl::get<UniquePtr<can::ICanTransport>>(std::move(maybe_transport));
+        EXPECT_THAT(maybe_transport, VariantWith<UniquePtr<ICanTransport>>(NotNull()));
+        auto transport = cetl::get<UniquePtr<ICanTransport>>(std::move(maybe_transport));
 
         transport->setLocalNodeId(local_node_id);
 
@@ -91,7 +92,7 @@ protected:
     // NOLINTBEGIN
     libcyphal::VirtualTimeScheduler scheduler_{};
     TrackingMemoryResource          mr_;
-    StrictMock<can::MediaMock>      media_mock_{};
+    StrictMock<MediaMock>           media_mock_{};
     // NOLINTEND
 };
 
@@ -166,11 +167,11 @@ TEST_F(TestCanSvcTxSessions, send_request)
     EXPECT_CALL(media_mock_, push(_, _, _)).WillOnce([&](auto deadline, auto can_id, auto payload) {
         EXPECT_THAT(now(), send_time + 10ms);
         EXPECT_THAT(deadline, send_time + timeout);
-        EXPECT_THAT(can_id, can::ServiceOfCanIdEq(123));
-        EXPECT_THAT(can_id, AllOf(can::SourceNodeOfCanIdEq(13), can::DestinationNodeOfCanIdEq(31)));
-        EXPECT_THAT(can_id, AllOf(can::PriorityOfCanIdEq(metadata.priority), can::IsServiceCanId()));
+        EXPECT_THAT(can_id, ServiceOfCanIdEq(123));
+        EXPECT_THAT(can_id, AllOf(SourceNodeOfCanIdEq(13), DestinationNodeOfCanIdEq(31)));
+        EXPECT_THAT(can_id, AllOf(PriorityOfCanIdEq(metadata.priority), IsServiceCanId()));
 
-        auto tbm = can::TailByteEq(metadata.transfer_id);
+        auto tbm = TailByteEq(metadata.transfer_id);
         EXPECT_THAT(payload, ElementsAre(tbm));
         return true;
     });
@@ -185,10 +186,10 @@ TEST_F(TestCanSvcTxSessions, send_request_with_argument_error)
 
     // Make initially anonymous node transport.
     //
-    std::array<can::IMedia*, 1> media_array{&media_mock_};
-    auto                        maybe_transport = can::makeTransport(mr_, media_array, 2);
-    ASSERT_THAT(maybe_transport, VariantWith<UniquePtr<can::ICanTransport>>(NotNull()));
-    auto transport = cetl::get<UniquePtr<can::ICanTransport>>(std::move(maybe_transport));
+    std::array<IMedia*, 1> media_array{&media_mock_};
+    auto                   maybe_transport = can::makeTransport(mr_, media_array, 2);
+    ASSERT_THAT(maybe_transport, VariantWith<UniquePtr<ICanTransport>>(NotNull()));
+    auto transport = cetl::get<UniquePtr<ICanTransport>>(std::move(maybe_transport));
 
     auto maybe_session = transport->makeRequestTxSession({123, 31});
     ASSERT_THAT(maybe_session, VariantWith<UniquePtr<IRequestTxSession>>(NotNull()));
@@ -224,11 +225,11 @@ TEST_F(TestCanSvcTxSessions, send_request_with_argument_error)
         EXPECT_CALL(media_mock_, push(_, _, _)).WillOnce([&](auto deadline, auto can_id, auto payload) {
             EXPECT_THAT(now(), send_time + 10ms);
             EXPECT_THAT(deadline, transfer_time + timeout);
-            EXPECT_THAT(can_id, can::ServiceOfCanIdEq(123));
-            EXPECT_THAT(can_id, AllOf(can::SourceNodeOfCanIdEq(13), can::DestinationNodeOfCanIdEq(31)));
-            EXPECT_THAT(can_id, AllOf(can::PriorityOfCanIdEq(metadata.priority), can::IsServiceCanId()));
+            EXPECT_THAT(can_id, ServiceOfCanIdEq(123));
+            EXPECT_THAT(can_id, AllOf(SourceNodeOfCanIdEq(13), DestinationNodeOfCanIdEq(31)));
+            EXPECT_THAT(can_id, AllOf(PriorityOfCanIdEq(metadata.priority), IsServiceCanId()));
 
-            auto tbm = can::TailByteEq(metadata.transfer_id);
+            auto tbm = TailByteEq(metadata.transfer_id);
             EXPECT_THAT(payload, ElementsAre(tbm));
             return true;
         });
@@ -298,11 +299,11 @@ TEST_F(TestCanSvcTxSessions, send_respose)
     EXPECT_CALL(media_mock_, push(_, _, _)).WillOnce([&](auto deadline, auto can_id, auto payload) {
         EXPECT_THAT(now(), send_time + 10ms);
         EXPECT_THAT(deadline, send_time + timeout);
-        EXPECT_THAT(can_id, can::ServiceOfCanIdEq(123));
-        EXPECT_THAT(can_id, AllOf(can::SourceNodeOfCanIdEq(31), can::DestinationNodeOfCanIdEq(13)));
-        EXPECT_THAT(can_id, AllOf(can::PriorityOfCanIdEq(metadata.priority), can::IsServiceCanId()));
+        EXPECT_THAT(can_id, ServiceOfCanIdEq(123));
+        EXPECT_THAT(can_id, AllOf(SourceNodeOfCanIdEq(31), DestinationNodeOfCanIdEq(13)));
+        EXPECT_THAT(can_id, AllOf(PriorityOfCanIdEq(metadata.priority), IsServiceCanId()));
 
-        auto tbm = can::TailByteEq(metadata.transfer_id);
+        auto tbm = TailByteEq(metadata.transfer_id);
         EXPECT_THAT(payload, ElementsAre(tbm));
         return true;
     });
@@ -317,10 +318,10 @@ TEST_F(TestCanSvcTxSessions, send_respose_with_argument_error)
 
     // Make initially anonymous node transport.
     //
-    std::array<can::IMedia*, 1> media_array{&media_mock_};
-    auto                        maybe_transport = can::makeTransport(mr_, media_array, 2);
-    ASSERT_THAT(maybe_transport, VariantWith<UniquePtr<can::ICanTransport>>(NotNull()));
-    auto transport = cetl::get<UniquePtr<can::ICanTransport>>(std::move(maybe_transport));
+    std::array<IMedia*, 1> media_array{&media_mock_};
+    auto                   maybe_transport = can::makeTransport(mr_, media_array, 2);
+    ASSERT_THAT(maybe_transport, VariantWith<UniquePtr<ICanTransport>>(NotNull()));
+    auto transport = cetl::get<UniquePtr<ICanTransport>>(std::move(maybe_transport));
 
     auto maybe_session = transport->makeResponseTxSession({123});
     ASSERT_THAT(maybe_session, VariantWith<UniquePtr<IResponseTxSession>>(NotNull()));
